Try to fix problem with getValue on forms
This might not be the right fix, but for list editors, it's important
that we use the getValue on the editor (to get a changeset and not
just an array) and not what's saved in newValues on the form. One
example of where this turns out to be a problem is in the attachments
on the wiki.